+++ src/app/top-toolbar/token/token.component.ts
@@ -1,6 +1,7 @@
-import {Component, OnInit, ViewChild} from '@angular/core';
-import {NgForm} from '@angular/forms';
+import {Component} from '@angular/core';
+import {FormControl, FormGroup, Validators} from '@angular/forms';
 import {PalladiumApiService} from '../../../services/palladium-api.service';
+import {MatDialog} from '@angular/material';
 
 
 @Component({
@@ -8,26 +9,37 @@ import {PalladiumApiService} from '../../../services/palladium-api.service';
   templateUrl: './token.component.html',
   styleUrls: ['./token.component.scss'],
 })
-export class TokenComponent implements OnInit {
-  tokens;
-  creating = false;
-  @ViewChild('Modal') Modal;
+export class TokenComponent {
 
-  constructor(private ApiService: PalladiumApiService) { }
-
-  ngOnInit() {}
+  constructor(public dialog: MatDialog) { }
 
   open() {
-    this.Modal.open();
+    this.dialog.open(TokenDialogComponent);
+  }
+}
+@Component({
+  selector: 'app-token-dialog',
+  templateUrl: 'app-token-dialog.html',
+})
+
+export class TokenDialogComponent {
+  tokens = [];
+  token_form = new FormGroup({
+    name: new FormControl('',  [Validators.required])
+  });
+
+  constructor(private ApiService: PalladiumApiService) {
     this.ApiService.get_tokens().then(tokens => {
       this.tokens = tokens;
     });
-
   }
-  create_token(form: NgForm) {
-    this.ApiService.create_token(form.value).then(token => {
+
+  get name() { return this.token_form.get('name'); }
+
+  create_token() {
+    console.log(this.name.value);
+    this.ApiService.create_token(this.name.value).then(token => {
       this.tokens.push(token['token_data']);
-      this.creating = false;
     });
   }
 }